[REFACTOR] Move api to shared folder
