+++ frontend/src/shared/api/reducer.js
