+++ frontend/src/shared/index.js
@@ -1,4 +1,5 @@
 import pt from "./i18n/pt.json";
 import en from "./i18n/en.json";
+import * as api from "./api";
 
-export { pt, en };
+export { pt, en, api };
